+++ data-manager/api/src/test/java/org/orbisgis/orbisdata/datamanager/api/dataset/IJdbcTableTest.java
@@ -314,6 +314,12 @@ public class IJdbcTableTest {
             return null;
         }
 
+        @NotNull
+        @Override
+        public Map<String, Object> firstRow() {
+            return null;
+        }
+
         @Override
         public boolean next() throws SQLException {
             if (!sqlException) {